reservationService/requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.0
psycopg2==2.9.9
python-dotenv==1.0.1
